Started scaffolding for User vs NonUser profile views
