+++ spoti-friends/src/common/views/AuthenticatedView.swift
@@ -3,12 +3,14 @@ import SwiftUI
 /// The view for when a user is signed into the app.
 struct AuthenticatedView: View {
     @StateObject var friendActivityViewModel: FriendActivityViewModel
+    @StateObject private var profileViewModel: ProfileViewModel
     @EnvironmentObject var authorizationViewModel: AuthorizationViewModel
     
     init() {
         _friendActivityViewModel = StateObject(
             wrappedValue: FriendActivityViewModel(user: nil, friendActivites: [])
         )
+        _profileViewModel = StateObject(wrappedValue: ProfileViewModel(user: AuthorizationViewModel().user))
         
         let standardAppearance = UITabBarAppearance()
         standardAppearance.backgroundColor = UIColor(Color.PresetColour.darkgrey)
@@ -25,11 +27,13 @@ struct AuthenticatedView: View {
                 Label("Friend Activity", systemImage: "figure.socialdance")
             }
             .environmentObject(friendActivityViewModel)
-            ProfileView(profile: friendActivityViewModel.user?.spotifyProfile ?? SpotifyProfileMock.jimHalpert).tabItem {
+            .environmentObject(profileViewModel)
+            UserProfileView(profile: friendActivityViewModel.user?.spotifyProfile ?? SpotifyProfileMock.jimHalpert).tabItem {
                 Label("My Profile", systemImage: "person")
             }
             .environmentObject(authorizationViewModel)
             .environmentObject(friendActivityViewModel)
+            .environmentObject(profileViewModel)
         }
         .tint(Color.PresetColour.spotifyGreen)
         .onAppear {
