+++ spoti-friends/src/profile/views/UserProfileView.swift
@@ -1,0 +1,158 @@
+import SwiftUI
+
+/// Renders the View for a user's profile, such as their details and top songs.
+///
+/// - Parameters:
+///   - profile: The Spotify profile to display.
+///   - recentTracks: A list of the profile's recently played tracks.
+///   - topTracks: A list of the profile's top tracks over the last month.
+///   - topArtists: A list of the profile's top artists over the last month.
+///
+/// - Returns: A view for the user's Spotify Profile.
+struct UserProfileView: View {
+    let profile: SpotifyProfile
+    @State private var recentTracks: ProfileViewModel.TracksWithResponseMetadata
+    @State private var topTracks: ProfileViewModel.TracksWithResponseMetadata
+    @State private var topArtists: ProfileViewModel.ArtistsWithResponseMetadata
+    @EnvironmentObject var profileViewModel: ProfileViewModel
+    @EnvironmentObject var authorizationViewModel: AuthorizationViewModel
+    @EnvironmentObject var friendActivityViewModel: FriendActivityViewModel
+    
+    /// NOTE: `topTracks` and `topArtists` default to `[]` so that we don't need to pass them in from `AuthenticatedView`.
+    /// However, we do want to pass in mock data for previews, so that's why we want them as optional parameters.
+    init(profile: SpotifyProfile, recentTracks: [Track] = [], topTracks: [Track] = [], topArtists: [Artist] = []) {
+        self.profile = profile
+        self.recentTracks = ProfileViewModel.TracksWithResponseMetadata(tracks: recentTracks)
+        self.topTracks = ProfileViewModel.TracksWithResponseMetadata(tracks: topTracks)
+        self.topArtists = ProfileViewModel.ArtistsWithResponseMetadata(artists: topArtists)
+    }
+    
+    var body: some View {
+        NavigationStack {
+            GeometryReader { reader in
+                ScrollView {
+                    VStack (alignment: .leading, spacing: 34) {
+                        // Profile Details
+                        ProfileDetails(profile: profile)
+                            .environmentObject(profileViewModel)
+                            .environmentObject(friendActivityViewModel)
+                        
+                        // Recent Tracks
+                        VStack (alignment: .leading) {
+                            Text("Recent Songs")
+                                .font(.body)
+                                .foregroundStyle(Color.PresetColour.whitePrimary)
+                            
+                            if recentTracks.isEmpty {
+                                Text("It's oddly quiet here...")
+                                    .font(.callout)
+                                    .foregroundStyle(Color.PresetColour.whiteSecondary)
+                                    .padding(.vertical, 4)
+                            } else {
+                                TrackList(tracks: recentTracks.tracks)
+                                ViewMoreButton(destination: ViewMoreRecentTracks(profile: profile)
+                                    .environmentObject(profileViewModel))
+                                .padding(.top, 4)
+                            }
+                        }
+                        
+                        // Top Tracks
+                        VStack (alignment: .leading) {
+                            Text("Top Songs")
+                                .font(.body)
+                                .foregroundStyle(Color.PresetColour.whitePrimary)
+                            Text("From this month")
+                                .font(.footnote)
+                                .foregroundStyle(Color.PresetColour.whiteSecondary)
+                            
+                            if topTracks.isEmpty {
+                                Text("Not enough listening data this month.")
+                                    .font(.callout)
+                                    .foregroundStyle(Color.PresetColour.whiteSecondary)
+                                    .padding(.vertical, 4)
+                            } else {
+                                TrackList(tracks: topTracks.tracks)
+                                ViewMoreButton(destination: ViewMoreTopSongs(profile: profile)
+                                    .environmentObject(profileViewModel))
+                                .padding(.top, 4)
+                            }
+                        }
+                        
+                        // Top Artists
+                        VStack (alignment: .leading) {
+                            Text("Top Artists")
+                                .font(.body)
+                                .foregroundStyle(Color.PresetColour.whitePrimary)
+                            Text("From this month")
+                                .font(.footnote)
+                                .foregroundStyle(Color.PresetColour.whiteSecondary)
+                            
+                            if topArtists.isEmpty {
+                                Text("Not enough listening data this month.")
+                                    .font(.callout)
+                                    .foregroundStyle(Color.PresetColour.whiteSecondary)
+                                    .padding(.vertical, 4)
+                            } else {
+                                ArtistList(artists: topArtists.artists)
+                                ViewMoreButton(destination: ViewMoreTopArtists(profile: profile)
+                                    .environmentObject(profileViewModel))
+                                .padding(.top, 4)
+                            }
+                        }
+                        Spacer()
+                    }
+                    .frame(minHeight: reader.size.height)
+                    .padding(.horizontal, 20)
+                    
+                    // Logout Button
+                    LogoutButton()
+                        .padding(.bottom, 10)
+                }
+                .padding(.top)
+                .background(Color.PresetGradient.profileViewGradient(profile: profile))
+                .onAppear {
+                    profileViewModel.user = authorizationViewModel.user
+                    
+                    Task {
+                        // Comment out these lines for SwiftUI Previews
+                        recentTracks = await profileViewModel.getRecentTracks(forProfile: profile, limit: 5) ?? ProfileViewModel.TracksWithResponseMetadata(tracks: [])
+                        topTracks = await profileViewModel.getTopTracks(forProfile: profile, timeRange: .oneMonth, limit: 5) ?? ProfileViewModel.TracksWithResponseMetadata(tracks: [])
+                        topArtists = await profileViewModel.getTopArtists(forProfile: profile, timeRange: .oneMonth, limit: 5) ?? ProfileViewModel.ArtistsWithResponseMetadata(artists: [])
+                    }
+                }
+            }
+        }
+    }
+    
+    // Logout button
+    struct LogoutButton: View {
+        @EnvironmentObject private var authorizationViewModel: AuthorizationViewModel
+        
+        var body: some View {
+            let buttonLabel = "Log out"
+            Button(action: {
+                authorizationViewModel.signOutUser()
+            }) {
+                Text(buttonLabel)
+                    .frame(width: 320, height: 50)
+                    .background(Color.PresetColour.transparentMaroon)
+                    .foregroundColor(Color.PresetColour.red)
+                    .fontWeight(.semibold)
+                    .cornerRadius(12)
+            }
+        }
+    }
+}
+
+#Preview {
+    ZStack {
+        let user = UserMock.userJimHalpert
+        let recentTracks = [TrackMock.iRememberEverything, TrackMock.luxury, TrackMock.traitor]
+        let topTracks = [TrackMock.iRememberEverything, TrackMock.luxury, TrackMock.traitor]
+        let topArtists = [ArtistMock.zachBryan, ArtistMock.jonBellion, ArtistMock.oliviaRodrigo, ArtistMock.kaceyMusgraves]
+        
+        UserProfileView(profile: user.spotifyProfile, recentTracks: recentTracks, topTracks: topTracks, topArtists: topArtists)
+            .environmentObject(AuthorizationViewModel())
+            .environmentObject(FriendActivityViewModel(user: user, friendActivites: []))
+    }
+}
